requirements.txt
flask==3.0.3
flask-cors==4.0.1
bcrypt==4.1.3
pyjwt==2.8.0
